refactor(workspace): move exec helpers into pkg/workspace for reuse
Moves FindRunningContainer, LoadExecResult, ResolveExecWorkdir,
ResolveDockerCommand, BuildExecEnv, ProbeContainerEnv, ContainerTarget,
DefaultDockerCommand, and ContainerStatusRunning out of cmd/workspace
into a new pkg/workspace/exec_helpers.go so the MCP server can call
them without depending on the cmd tree.

Re-exports are provided in cmd/workspace/exec.go for backward
compatibility with existing callers in cmd/config and cmd/internal.
