+++ cmd/internal/runusercommands.go
@@ -233,7 +233,7 @@ func (cmd *RunUserCommandsCmd) resolveDockerPath() string {
 	if cmd.DockerPath != "" {
 		return cmd.DockerPath
 	}
-	return workspace.DefaultDockerCommand
+	return workspace2.DefaultDockerCommand
 }
 
 func (cmd *RunUserCommandsCmd) inspectRunningContainer(
@@ -252,7 +252,7 @@ func (cmd *RunUserCommandsCmd) inspectRunningContainer(
 	}
 
 	containerDetails := &details[0]
-	if !strings.EqualFold(containerDetails.State.Status, workspace.ContainerStatusRunning) {
+	if !strings.EqualFold(containerDetails.State.Status, workspace2.ContainerStatusRunning) {
 		errMsg := fmt.Sprintf(
 			"container %s is not running (status: %s)",
 			cmd.ContainerID,
@@ -332,12 +332,12 @@ func (cmd *RunUserCommandsCmd) resolveContainer(
 	}
 
 	workspaceConfig := client.WorkspaceConfig()
-	dockerCommand := workspace.ResolveDockerCommand(workspaceConfig)
+	dockerCommand := workspace2.ResolveDockerCommand(workspaceConfig)
 	if cmd.DockerPath != "" {
 		dockerCommand = cmd.DockerPath
 	}
 
-	containerDetails, err := workspace.FindRunningContainer(
+	containerDetails, err := workspace2.FindRunningContainer(
 		ctx, dockerCommand, devcontainer.GetRunnerIDFromWorkspace(workspaceConfig), cmd.IDLabels,
 	)
 	if err != nil {
@@ -345,7 +345,7 @@ func (cmd *RunUserCommandsCmd) resolveContainer(
 		return nil, nil, err
 	}
 
-	result := workspace.LoadExecResult(workspaceConfig, containerDetails)
+	result := workspace2.LoadExecResult(workspaceConfig, containerDetails)
 	if result == nil || result.MergedConfig == nil {
 		_ = devcconfig.WriteErrorJSON(
 			os.Stderr,
@@ -372,7 +372,7 @@ func (cmd *RunUserCommandsCmd) resolveContainer(
 		Helper:      &docker.DockerHelper{DockerCommand: dockerCommand},
 		ContainerID: containerDetails.ID,
 		EnvArgs:     envArgs,
-		Workdir:     workspace.ResolveExecWorkdir(result, client.Workspace()),
+		Workdir:     workspace2.ResolveExecWorkdir(result, client.Workspace()),
 		User:        devcconfig.GetRemoteUser(result),
 	}
 	return params, result, nil
